Changes to support renaming the anaytics-common repo to shared-analytics.
