+++ components/log-analyzer/org.wso2.carbon.shared.analytics.la.apim/src/main/java/org/wso2/carbon/shared/analytics/la/apim/internal/ServiceReferenceHolder.java
@@ -1,0 +1,42 @@
+/*
+ * Copyright (c) 2016, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ *
+ * WSO2 Inc. licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package org.wso2.carbon.shared.analytics.la.apim.internal;
+
+import org.wso2.carbon.apimgt.impl.APIManagerConfigurationService;
+
+/**
+ * Singleton service reference holder bridging OSGI component to other java classes.
+ */
+public class ServiceReferenceHolder {
+
+    private static ServiceReferenceHolder instance = new ServiceReferenceHolder();
+    private APIManagerConfigurationService apiManagerConfigurationService;
+
+    public APIManagerConfigurationService getAPIManagerConfigurationService() {
+        return apiManagerConfigurationService;
+    }
+
+    public void setAPIManagerConfigurationService(APIManagerConfigurationService apiManagerConfigurationService) {
+        this.apiManagerConfigurationService = apiManagerConfigurationService;
+    }
+
+    public static ServiceReferenceHolder getInstance() {
+        return instance;
+    }
+}
