+++ components/log4j/org.wso2.shared.analytics.data.agents.log4j/src/main/java/org/wso2/carbon/data/agents/log4j/util/TenantAwarePatternLayout.java
