--- components/util/org.wso2.analytics.common.util/src/main/java/org/wso2/analytics/common/util/time/bean/TimeRange.java
@@ -1,49 +1,0 @@
-/*
-* Copyright (c) 2016, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
-*
-* WSO2 Inc. licenses this file to you under the Apache License,
-* Version 2.0 (the "License"); you may not use this file except
-* in compliance with the License.
-* You may obtain a copy of the License at
-*
-* http://www.apache.org/licenses/LICENSE-2.0
-*
-* Unless required by applicable law or agreed to in writing,
-* software distributed under the License is distributed on an
-* "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-* KIND, either express or implied.  See the License for the
-* specific language governing permissions and limitations
-* under the License.
-*/
-
-package org.wso2.analytics.common.util.time.bean;
-
-public class TimeRange {
-
-    private String unit;
-    private long[] range;
-
-    public String getUnit() {
-        return unit;
-    }
-
-    public void setUnit(String unit) {
-        this.unit = unit;
-    }
-
-    public long[] getRange() {
-        return range;
-    }
-
-    public void setRange(long[] range) {
-        this.range = range;
-    }
-
-    public TimeRange() {
-    }
-
-    public TimeRange(String unit, long[] range) {
-        this.unit = unit;
-        this.range = range;
-    }
-}
