+++ features/filetail/org.wso2.shared.analytics.data.agents.filetail.feature/pom.xml
@@ -1,0 +1,89 @@
+<?xml version="1.0" encoding="utf-8"?>
+
+<!--
+  ~ Copyright (c) 2016, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+  ~
+  ~ WSO2 Inc. licenses this file to you under the Apache License,
+  ~ Version 2.0 (the "License"); you may not use this file except
+  ~ in compliance with the License.
+  ~ You may obtain a copy of the License at
+  ~
+  ~ http://www.apache.org/licenses/LICENSE-2.0
+  ~
+  ~ Unless required by applicable law or agreed to in writing,
+  ~ software distributed under the License is distributed on an
+  ~ "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+  ~ KIND, either express or implied.  See the License for the
+  ~ specific language governing permissions and limitations
+  ~ under the License.
+  -->
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
+
+    <parent>
+        <groupId>org.wso2.shared.analytics</groupId>
+        <artifactId>filetail-feature</artifactId>
+        <version>1.0.1-SNAPSHOT</version>
+        <relativePath>../pom.xml</relativePath>
+    </parent>
+    <modelVersion>4.0.0</modelVersion>
+
+    <name>WSO2 - Analytics File Tail Feature</name>
+    <artifactId>org.wso2.shared.analytics.data.agents.filetail.feature</artifactId>
+    <packaging>pom</packaging>
+
+    <build>
+  		<plugins>
+			<plugin>
+				<artifactId>maven-resources-plugin</artifactId>
+				<executions>
+					<execution>
+						<id>copy-resources</id>
+						<phase>generate-resources</phase>
+						<goals>
+							<goal>copy-resources</goal>
+						</goals>
+						<configuration>
+							<outputDirectory>src/main/resources</outputDirectory>
+							<resources>
+								<resource>
+									<directory>resources</directory>
+									<includes>
+										<include>p2.inf</include>
+										<include>build.properties</include>
+									</includes>
+								</resource>
+							</resources>
+						</configuration>
+					</execution>
+				</executions>
+			</plugin>
+  			<plugin>
+  				<groupId>org.wso2.maven</groupId>
+  				<artifactId>carbon-p2-plugin</artifactId>
+  				<version>1.5.3</version>
+  				<executions>
+  					<execution>
+  						<id>p2-feature-generation</id>
+  						<phase>package</phase>
+  						<goals>
+  							<goal>p2-feature-gen</goal>
+  						</goals>
+  						<configuration>
+  							<id>org.wso2.shared.analytics.data.agents.filetail</id>
+  							<propertiesFile>../../etc/feature.properties</propertiesFile>
+  							<adviceFile>
+  								<properties>
+  									<propertyDef>org.wso2.carbon.p2.category.type:server</propertyDef>
+                    							<propertyDef>org.eclipse.equinox.p2.type.group:true</propertyDef>
+  								</properties>
+  							</adviceFile>
+							<bundles>
+								<bundleDef>org.wso2.shared.analytics:org.wso2.shared.analytics.data.agents.filetail:${pom.version}</bundleDef>
+							</bundles>
+  						</configuration>
+  					</execution>
+  				</executions>
+  			</plugin>
+  		</plugins>
+  	</build>
+</project>
