+++ pom.xml
@@ -20,11 +20,11 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
 
 	<modelVersion>4.0.0</modelVersion>
-	<groupId>org.wso2.analytics.common</groupId>
-	<artifactId>analytics-common</artifactId>
+	<groupId>org.wso2.shared.analytics</groupId>
+	<artifactId>shared-analytics</artifactId>
 	<packaging>pom</packaging>
 	<version>1.0.1-SNAPSHOT</version>
-	<name>WSO2 Analytics Common - Aggregator Module</name>
+	<name>WSO2 Shared Analytics - Aggregator Module</name>
 	<url>http://wso2.com/products/api-manager</url>
 	<description>WSO2 Analytics for API Manager</description>
 
@@ -35,9 +35,9 @@
 	</parent>
 
 	<scm>
-		<url>https://github.com/wso2/analytics-common.git</url>
-		<developerConnection>scm:git:https://github.com/wso2/analytics-common.git</developerConnection>
-		<connection>scm:git:https://github.com/wso2/analytics-common.git</connection>
+		<url>https://github.com/wso2/shared-analytics.git</url>
+		<developerConnection>scm:git:https://github.com/wso2/shared-analytics.git</developerConnection>
+		<connection>scm:git:https://github.com/wso2/shared-analytics.git</connection>
 		<tag>HEAD</tag>
 	</scm>
